ImprovedLoginFuncAndCreatedAddressRepository

+++ STMDomain/Domain/FamilyData.cs
@@ -2,7 +2,7 @@
 {
     public class FamilyData : Entity
     {
-        public float FamilyRevenue { get; set; }
+        public decimal FamilyRevenue { get; set; }
         public int ResidenceQuantity { get; set; }
         public int ChildrenQuantity { get; set; }
         public bool IsSoloParent { get; set; } = false;
